Introduce a soft token section and add WearAuthn (#25)
Introduces a section for soft tokens, which fit neither into the existing hardware nor the client lib section. Damian Czaja's android-webauthn-token is an example of such a soft token, this commit moves it into this new section.

It also adds WearAuthn to the list of soft tokens, which I developed to provide a full-fledged open-source implementation of the CTAP2 standard in a higher programming language (Kotlin/Java). It uses the watch's screen to show confirmation prompts and supports the on-device selection of resident credentials.