Update README.md
Added a Cheat Sheet for passkey implementation to the section "Other".
The Cheat Sheet is free and explains all relevant objects, concepts, methods, ... from WebAuthn. 
It serves as a good directory for developers to find specific information instead of going through the WebAuthn specification.

Also removed a typo in a different section.